+++ packages/rum-core/src/browser/windowOpenObservable.ts
@@ -1,0 +1,7 @@
+import { instrumentMethod, Observable } from '@datadog/browser-core'
+
+export function createWindowOpenObservable() {
+  const observable = new Observable<void>()
+  const { stop } = instrumentMethod(window, 'open', () => observable.notify())
+  return { observable, stop }
+}