+++ packages/rum-core/src/domain/configuration/configuration.ts
@@ -65,7 +65,7 @@ export interface RumInitConfiguration extends InitConfiguration {
   traceSampleRate?: number | undefined
   /**
    * If you set a `traceSampleRate`, to ensure backend services' sampling decisions are still applied, configure the `traceContextInjection` initialization parameter to sampled.
-   * @default all
+   * @default sampled
    * See [Connect RUM and Traces](https://docs.datadoghq.com/real_user_monitoring/platform/connect_rum_and_traces/?tab=browserrum) for further information.
    */
   traceContextInjection?: TraceContextInjection | undefined
@@ -100,6 +100,7 @@ export interface RumInitConfiguration extends InitConfiguration {
   /**
    * Enables automatic collection of users actions.
    * See [Tracking User Actions](https://docs.datadoghq.com/real_user_monitoring/browser/tracking_user_actions) for further information.
+   * @default true
    */
   trackUserInteractions?: boolean | undefined
   /**
@@ -115,10 +116,12 @@ export interface RumInitConfiguration extends InitConfiguration {
   trackViewsManually?: boolean | undefined
   /**
    * Enables collection of resource events.
+   * @default true
    */
   trackResources?: boolean | undefined
   /**
    * Enables collection of long task events.
+   * @default true
    */
   trackLongTasks?: boolean | undefined
 
@@ -202,10 +205,10 @@ export function validateAndBuildRumConfiguration(
     excludedActivityUrls: initConfiguration.excludedActivityUrls ?? [],
     workerUrl: initConfiguration.workerUrl,
     compressIntakeRequests: !!initConfiguration.compressIntakeRequests,
-    trackUserInteractions: !!initConfiguration.trackUserInteractions,
+    trackUserInteractions: !!(initConfiguration.trackUserInteractions ?? true),
     trackViewsManually: !!initConfiguration.trackViewsManually,
-    trackResources: !!initConfiguration.trackResources,
-    trackLongTasks: !!initConfiguration.trackLongTasks,
+    trackResources: !!(initConfiguration.trackResources ?? true),
+    trackLongTasks: !!(initConfiguration.trackLongTasks ?? true),
     subdomain: initConfiguration.subdomain,
     defaultPrivacyLevel: objectHasValue(DefaultPrivacyLevel, initConfiguration.defaultPrivacyLevel)
       ? initConfiguration.defaultPrivacyLevel
@@ -214,7 +217,7 @@ export function validateAndBuildRumConfiguration(
     customerDataTelemetrySampleRate: 1,
     traceContextInjection: objectHasValue(TraceContextInjection, initConfiguration.traceContextInjection)
       ? initConfiguration.traceContextInjection
-      : TraceContextInjection.ALL,
+      : TraceContextInjection.SAMPLED,
     plugins: initConfiguration.plugins || [],
     ...baseConfiguration,
   }