+++ packages/rum-core/src/domain/tracing/tracer.spec.ts
@@ -4,7 +4,9 @@ import { createRumSessionManagerMock } from '../../../test'
 import type { RumFetchResolveContext, RumFetchStartContext, RumXhrStartContext } from '../requestCollection'
 import type { RumConfiguration, RumInitConfiguration } from '../configuration'
 import { validateAndBuildRumConfiguration } from '../configuration'
-import { startTracer, createTraceIdentifier, type TraceIdentifier, getCrypto } from './tracer'
+import { startTracer } from './tracer'
+import type { SpanIdentifier, TraceIdentifier } from './identifier'
+import { createSpanIdentifier, createTraceIdentifier } from './identifier'
 
 describe('tracer', () => {
   let configuration: RumConfiguration
@@ -75,8 +77,7 @@ describe('tracer', () => {
     })
 
     it("should trace request with priority '1' when sampled", () => {
-      spyOn(Math, 'random').and.callFake(() => 0)
-      const tracer = startTracer({ ...configuration, traceSampleRate: 50 }, sessionManager)
+      const tracer = startTracer({ ...configuration, traceSampleRate: 100 }, sessionManager)
       const context = { ...ALLOWED_DOMAIN_CONTEXT }
       tracer.traceXhr(context, xhr as unknown as XMLHttpRequest)
 
@@ -86,9 +87,11 @@ describe('tracer', () => {
       expect(xhr.headers).toEqual(tracingHeadersFor(context.traceId!, context.spanId!, '1'))
     })
 
-    it("should trace request with priority '0' when not sampled", () => {
-      spyOn(Math, 'random').and.callFake(() => 1)
-      const tracer = startTracer({ ...configuration, traceSampleRate: 50 }, sessionManager)
+    it("should trace request with priority '0' when not sampled and config set to all", () => {
+      const tracer = startTracer(
+        { ...configuration, traceSampleRate: 0, traceContextInjection: TraceContextInjection.ALL },
+        sessionManager
+      )
       const context = { ...ALLOWED_DOMAIN_CONTEXT }
       tracer.traceXhr(context, xhr as unknown as XMLHttpRequest)
 
@@ -98,12 +101,11 @@ describe('tracer', () => {
       expect(xhr.headers).toEqual(tracingHeadersFor(context.traceId!, context.spanId!, '0'))
     })
 
-    it("should trace request with sampled set to '0' in OTel headers when not sampled", () => {
-      spyOn(Math, 'random').and.callFake(() => 1)
-
+    it("should trace request with sampled set to '0' in OTel headers when not sampled and config set to all", () => {
       const configurationWithAllOtelHeaders = validateAndBuildRumConfiguration({
         ...INIT_CONFIGURATION,
-        traceSampleRate: 50,
+        traceSampleRate: 0,
+        traceContextInjection: TraceContextInjection.ALL,
         allowedTracingUrls: [{ match: window.location.origin, propagatorTypes: ['b3', 'tracecontext', 'b3multi'] }],
       })!
 
@@ -214,7 +216,6 @@ describe('tracer', () => {
       const configurationWithInjectionParam = {
         ...configuration,
         traceSampleRate: 0,
-        traceContextInjection: TraceContextInjection.SAMPLED,
       }
 
       const tracer = startTracer(configurationWithInjectionParam, sessionManager)
@@ -229,7 +230,6 @@ describe('tracer', () => {
       const configurationWithInjectionParam = {
         ...configuration,
         traceSampleRate: 100,
-        traceContextInjection: TraceContextInjection.SAMPLED,
       }
 
       const tracer = startTracer(configurationWithInjectionParam, sessionManager)
@@ -455,8 +455,7 @@ describe('tracer', () => {
     it("should trace request with priority '1' when sampled", () => {
       const context: Partial<RumFetchStartContext> = { ...ALLOWED_DOMAIN_CONTEXT }
 
-      spyOn(Math, 'random').and.callFake(() => 0)
-      const tracer = startTracer({ ...configuration, traceSampleRate: 50 }, sessionManager)
+      const tracer = startTracer({ ...configuration, traceSampleRate: 100 }, sessionManager)
       tracer.traceFetch(context)
 
       expect(context.traceSampled).toBe(true)
@@ -465,11 +464,13 @@ describe('tracer', () => {
       expect(context.init!.headers).toEqual(tracingHeadersAsArrayFor(context.traceId!, context.spanId!, '1'))
     })
 
-    it("should trace request with priority '0' when not sampled", () => {
+    it("should trace request with priority '0' when not sampled and config set to all", () => {
       const context: Partial<RumFetchStartContext> = { ...ALLOWED_DOMAIN_CONTEXT }
 
-      spyOn(Math, 'random').and.callFake(() => 1)
-      const tracer = startTracer({ ...configuration, traceSampleRate: 50 }, sessionManager)
+      const tracer = startTracer(
+        { ...configuration, traceSampleRate: 0, traceContextInjection: TraceContextInjection.ALL },
+        sessionManager
+      )
       tracer.traceFetch(context)
 
       expect(context.traceSampled).toBe(false)
@@ -622,9 +623,9 @@ describe('tracer', () => {
       const context: RumFetchResolveContext = {
         status: 0,
 
-        spanId: createTraceIdentifier(),
+        spanId: createSpanIdentifier(),
         traceId: createTraceIdentifier(),
-      } as any
+      } satisfies Partial<RumFetchResolveContext> as any
       tracer.clearTracingIfNeeded(context)
 
       expect(context.traceId).toBeUndefined()
@@ -636,30 +637,14 @@ describe('tracer', () => {
       const context: RumFetchResolveContext = {
         status: 200,
 
-        spanId: createTraceIdentifier(),
+        spanId: createSpanIdentifier(),
         traceId: createTraceIdentifier(),
-      } as any
+      } satisfies Partial<RumFetchResolveContext> as any
       tracer.clearTracingIfNeeded(context)
 
       expect(context.traceId).toBeDefined()
       expect(context.spanId).toBeDefined()
     })
-  })
-})
-
-describe('TraceIdentifier', () => {
-  it('should generate id', () => {
-    const identifier = createTraceIdentifier()
-
-    expect(identifier.toDecimalString()).toMatch(/^\d+$/)
-  })
-
-  it('should pad the string to 16 characters', () => {
-    spyOn(getCrypto() as any, 'getRandomValues').and.callFake((buffer: Uint8Array) => {
-      buffer[buffer.length - 1] = 0x01
-    })
-    const identifier = createTraceIdentifier()
-    expect(identifier.toPaddedHexadecimalString()).toEqual('0000000000000001')
   })
 })
 
@@ -671,16 +656,16 @@ function toPlainObject(headers: Headers) {
   return result
 }
 
-function tracingHeadersFor(traceId: TraceIdentifier, spanId: TraceIdentifier, samplingPriority: '1' | '0') {
+function tracingHeadersFor(traceId: TraceIdentifier, spanId: SpanIdentifier, samplingPriority: '1' | '0') {
   return {
     'x-datadog-origin': 'rum',
-    'x-datadog-parent-id': spanId.toDecimalString(),
+    'x-datadog-parent-id': spanId.toString(),
     'x-datadog-sampling-priority': samplingPriority,
-    'x-datadog-trace-id': traceId.toDecimalString(),
+    'x-datadog-trace-id': traceId.toString(),
   }
 }
 
-function tracingHeadersAsArrayFor(traceId: TraceIdentifier, spanId: TraceIdentifier, samplingPriority: '1' | '0') {
+function tracingHeadersAsArrayFor(traceId: TraceIdentifier, spanId: SpanIdentifier, samplingPriority: '1' | '0') {
   return objectEntries(tracingHeadersFor(traceId, spanId, samplingPriority))
 }